Reorganize compound components
